Crate sync_wrapper

A mutual exclusion primitive that relies on static type information only

This library is inspired by this discussion.

Structs

SyncFuture

Future which is Sync.

SyncStream

Stream which is Sync.

SyncWrapper

A mutual exclusion primitive that relies on static type information only